backend/requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-cors-headers==2.5.3
django-rest-swagger==2.2.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pygments==2.4.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.2
wrapt==1.11.1
